+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :feature:`769` Added a new configuration flag, ``log.show_raw_param_values``, defaulting to ``False``. If set to True, log lines for beginnings of tests will contain actual parametrization values instead of format-safe strings.
 * :feature:`528` ``slash.exclude`` can now exclude combinations of parameter values
 * :bug:`783` Session errors in children are now handled and reported when running with parallel
 * :feature:`785` Plugins can now be marked to indicate whether or not they support parallel